requirements.txt
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
textblob==0.17.1
nltk==3.8.1
Pillow==10.2.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
Django==5.0.1